alpr/requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
certifi==2019.11.28
chardet==3.0.4
cryptography==2.1.4
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.6
imageio==2.6.1
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.2.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
opencv-python==4.2.0.32
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.5.2
pygobject==3.26.1
pyrsistent==0.15.7
python-apt==1.6.4
python-dateutil==2.8.1
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2019.3
pyxdg==0.25
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.23.0
scikit-learn==0.22
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
six==1.13.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.6.0
torchvision==0.7.0
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
unattended-upgrades==0.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.0.0
